requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
PyYAML
jinja2
requests
google-genai